fix sim building with x11, address x11 rendering issues

+++ include/mui/tools.h
@@ -22,7 +22,7 @@ namespace mui
 
         CPUMonitorUsage();
 
-        float usage(unsigned int index = 0) const { return m_cpu_usage[index]; }
+        double usage(unsigned int index = 0) const { return m_cpu_usage[index]; }
 
         /**
          * Update the CPU usage.
@@ -34,9 +34,9 @@ namespace mui
 
     private:
 
-        float m_total_cpu_last_time[4] = {0};
-        float m_work_cpu_last_time[4] = {0};
-        float m_cpu_usage[4] = {0};
+        double m_total_cpu_last_time[4] = {0};
+        double m_work_cpu_last_time[4] = {0};
+        double m_cpu_usage[4] = {0};
     };
 
 }
